% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{codelist_extract_ids}
\alias{codelist_extract_ids}
\title{Extract codelist IDs}
\usage{
codelist_extract_ids(cl_df)
}
\arguments{
\item{cl_df}{A tibble returned by \code{\link[=get_codelists]{get_codelists()}}.}
}
\value{
A character vector of codelist IDs.
}
\description{
Extract codelist IDs
}
\examples{
\donttest{
scb <- px_api("scb", lang = "en")
if (px_available(scb)) {
  get_codelists(scb, "TAB638", "Region") |> codelist_extract_ids()
}}
}
