% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{codelist_values}
\alias{codelist_values}
\title{Extract values for a specific codelist}
\usage{
codelist_values(cl_df, codelist_id)
}
\arguments{
\item{cl_df}{A tibble returned by \code{\link[=get_codelists]{get_codelists()}}.}

\item{codelist_id}{Codelist ID (character).}
}
\value{
A tibble with columns \code{code} and \code{text}.
}
\description{
Extract values for a specific codelist
}
\examples{
\dontrun{
scb <- px_api("scb", lang = "en")
cls <- get_codelists(scb, "TAB638", "Region")
codelist_values(cls, cls$id[1])
}
}
