% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{compose_table_query}
\alias{compose_table_query}
\title{Compose a table query URL}
\usage{
compose_table_query(
  api,
  query = NULL,
  id = NULL,
  updated_since = NULL,
  page = NA,
  per_page = NA
)
}
\arguments{
\item{api}{A \verb{<px_api>} object.}

\item{query}{Free-text search string.}

\item{id}{Table ID(s).}

\item{updated_since}{Days since last update.}

\item{page}{Page number.}

\item{per_page}{Results per page.}
}
\value{
A character URL string.
}
\description{
Build the URL for querying the tables endpoint (advanced use).
}
\examples{
\donttest{
scb <- px_api("scb", lang = "en")
if (px_available(scb)) {
  compose_table_query(scb, query = "population")
}}
}
