% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_minimize}
\alias{data_minimize}
\title{Remove monotonous columns from a data tibble}
\usage{
data_minimize(data_df, remove_monotonous_data = TRUE)
}
\arguments{
\item{data_df}{A tibble returned by \code{\link[=get_data]{get_data()}}.}

\item{remove_monotonous_data}{Remove columns where all values are identical.}
}
\value{
A tibble.
}
\description{
Remove monotonous columns from a data tibble
}
\examples{
\donttest{
scb <- px_api("scb", lang = "en")
if (px_available(scb)) {
  d <- get_data(scb, "TAB638", Region = "0180", Tid = px_top(3))
  d |> data_minimize()
}}
}
