% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixieweb-package.R
\docType{package}
\name{pixieweb-package}
\alias{pixieweb}
\alias{pixieweb-package}
\title{pixieweb: Access PX-Web Statistical Data from R}
\description{
A pipe-friendly R client for PX-Web statistical APIs. Provides a search-then-fetch workflow for discovering and downloading data from national statistics agencies (SCB, SSB, Statistics Finland, etc.) using a consistent tibble-based interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lchansson.github.io/pixieweb}
  \item \url{https://github.com/lchansson/pixieweb}
  \item Report bugs at \url{https://github.com/lchansson/pixieweb/issues}
}

}
\author{
\strong{Maintainer}: Love Hansson \email{love.hansson@gmail.com} [copyright holder]

}
\keyword{internal}
