% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{px_available}
\alias{px_available}
\title{Check if a PX-Web API is reachable}
\usage{
px_available(api)
}
\arguments{
\item{api}{A \verb{<px_api>} object.}
}
\value{
Logical: \code{TRUE} if the API responds, \code{FALSE} otherwise.
}
\description{
Check if a PX-Web API is reachable
}
\examples{
\donttest{
scb <- px_api("scb")
if (px_available(scb)) {
  px_available(scb)
}}
}
