% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{px_cite}
\alias{px_cite}
\title{Generate a citation for downloaded data}
\usage{
px_cite(data_df)
}
\arguments{
\item{data_df}{A tibble returned by \code{\link[=get_data]{get_data()}}.}
}
\value{
A character string (formatted citation).
}
\description{
Produces a citation string from metadata attached to data by \code{\link[=get_data]{get_data()}}.
}
\examples{
\donttest{
scb <- px_api("scb", lang = "en")
if (px_available(scb)) {
  d <- get_data(scb, "TAB638", Region = "0180", Tid = px_top(3))
  px_cite(d)
}}
}
