% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{px_selections}
\alias{px_selections}
\alias{px_all}
\alias{px_top}
\alias{px_bottom}
\alias{px_from}
\alias{px_to}
\alias{px_range}
\alias{print.px_selection}
\title{Create a PX-Web selection object}
\usage{
px_all(pattern = "*")

px_top(n)

px_bottom(n)

px_from(value)

px_to(value)

px_range(from, to)

\method{print}{px_selection}(x, ...)
}
\arguments{
\item{pattern}{Glob pattern (default \code{"*"} for all).}

\item{n}{Number of values.}

\item{value}{Value code (inclusive bound).}

\item{from, to}{Value codes for range bounds (inclusive).}

\item{x}{A \verb{<px_selection>} object.}

\item{...}{Ignored.}
}
\value{
A \verb{<px_selection>} object.
}
\description{
These helpers create \verb{<px_selection>} objects that \code{get_data()} translates
into the appropriate API filter. Each represents a different way to select
variable values in PX-Web queries.
}
