% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_extract_ids}
\alias{table_extract_ids}
\title{Extract table IDs from a table tibble}
\usage{
table_extract_ids(table_df)
}
\arguments{
\item{table_df}{A tibble returned by \code{\link[=get_tables]{get_tables()}}.}
}
\value{
A character vector of table IDs.
}
\description{
Extract table IDs from a table tibble
}
\examples{
\dontrun{
scb <- px_api("scb", lang = "en")
get_tables(scb, query = "population") |> table_extract_ids()
}
}
