% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_minimize}
\alias{table_minimize}
\title{Remove monotonous columns from a table tibble}
\usage{
table_minimize(table_df, remove_monotonous_data = TRUE)
}
\arguments{
\item{table_df}{A tibble returned by \code{\link[=get_tables]{get_tables()}}.}

\item{remove_monotonous_data}{Remove columns where all values are identical.}
}
\value{
A tibble.
}
\description{
Remove monotonous columns from a table tibble
}
\examples{
\dontrun{
scb <- px_api("scb", lang = "en")
get_tables(scb, query = "population") |> table_minimize()
}
}
