% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variable_extract_ids}
\alias{variable_extract_ids}
\title{Extract variable codes from a variable tibble}
\usage{
variable_extract_ids(var_df)
}
\arguments{
\item{var_df}{A tibble returned by \code{\link[=get_variables]{get_variables()}}.}
}
\value{
A character vector of variable codes.
}
\description{
Extract variable codes from a variable tibble
}
\examples{
\dontrun{
scb <- px_api("scb", lang = "en")
get_variables(scb, "TAB638") |> variable_extract_ids()
}
}
