% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variable_values}
\alias{variable_values}
\title{Extract values for a specific variable}
\usage{
variable_values(var_df, variable_code)
}
\arguments{
\item{var_df}{A tibble returned by \code{\link[=get_variables]{get_variables()}}.}

\item{variable_code}{Variable code (character).}
}
\value{
A tibble with columns \code{code} and \code{text}.
}
\description{
Extract values for a specific variable
}
\examples{
\dontrun{
scb <- px_api("scb", lang = "en")
vars <- get_variables(scb, "TAB638")
vars |> variable_values("Kon")
}
}
