% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_legend}
\alias{data_legend}
\title{Generate a plot legend from variable metadata}
\usage{
data_legend(data_df, var_df)
}
\arguments{
\item{data_df}{A tibble returned by \code{\link[=get_data]{get_data()}}.}

\item{var_df}{A tibble returned by \code{\link[=get_variables]{get_variables()}}.}
}
\value{
A character string suitable for plot captions.
}
\description{
Generate a plot legend from variable metadata
}
\examples{
\donttest{
scb <- px_api("scb", lang = "en")
if (px_available(scb)) {
  vars <- get_variables(scb, "TAB638")
  d <- get_data(scb, "TAB638", Region = "0180", Tid = px_top(3))
  data_legend(d, vars)
}}
}
